Não pode ser excluída ou alterada uma eleição com votos. [FALTA TESTAR]

+++ src/main/java/br/edu/ulbra/election/election/service/ElectionService.java
@@ -7,6 +7,7 @@ import br.edu.ulbra.election.election.model.Election;
 import br.edu.ulbra.election.election.output.v1.ElectionOutput;
 import br.edu.ulbra.election.election.output.v1.GenericOutput;
 import br.edu.ulbra.election.election.repository.ElectionRepository;
+import br.edu.ulbra.election.election.repository.VoteRepository;
 import org.apache.commons.lang.StringUtils;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeToken;
@@ -21,11 +22,13 @@ public class ElectionService {
 
     private final ElectionRepository electionRepository;
     private final ModelMapper modelMapper;
+    private final VoteRepository voteRepository;
     
     @Autowired
-    public ElectionService(ElectionRepository electionRepository, ModelMapper modelMapper){
+    public ElectionService(ElectionRepository electionRepository, ModelMapper modelMapper, VoteRepository voteRepository){
         this.electionRepository = electionRepository;
         this.modelMapper = modelMapper;
+        this.voteRepository = voteRepository;
     }
 
 
@@ -85,6 +88,10 @@ public class ElectionService {
         Election election = electionRepository.findById(electionId).orElse(null);
         if (election == null){
             throw new GenericOutputException(MESSAGE_ELECTION_NOT_FOUND);
+        }
+
+        if(voteRepository.getVoteByElection(election) != null){ //retornando um voto pela elição
+            throw new GenericOutputException("Election already have votes");
         }
 
         electionRepository.delete(election);
